requirements.txt
django==4.1.1
pillow==9.2.0
django-cleanup==6.0.0
django-filter==22.1